+++ Sources/Swift/Helper/SentryFileManager.swift
@@ -97,7 +97,7 @@
     }
     
     @objc public func storeCurrentSession(_ session: SentrySession) {
-        helper.storeCurrentSessionData(SentrySerializationSwift.data(withJSONObject: session.serialize()))
+        helper.storeCurrentSessionData(SentrySerializationSwift.data(withJSONObject: session.serializeForPersistence()))
     }
     
     @objc public func readCurrentSession() -> SentrySession? {
@@ -109,7 +109,7 @@
     }
     
     @objc public func storeCrashedSession(_ session: SentrySession) {
-        helper.storeCrashedSessionData(SentrySerializationSwift.data(withJSONObject: session.serialize()))
+        helper.storeCrashedSessionData(SentrySerializationSwift.data(withJSONObject: session.serializeForPersistence()))
     }
 
     @objc public func readCrashedSession() -> SentrySession? {
@@ -121,7 +121,7 @@
     }
 
     @objc public func storeAbnormalSession(_ session: SentrySession) {
-        helper.storeAbnormalSessionData(SentrySerializationSwift.data(withJSONObject: session.serialize()))
+        helper.storeAbnormalSessionData(SentrySerializationSwift.data(withJSONObject: session.serializeForPersistence()))
     }
     
     @objc public func readAbnormalSession() -> SentrySession? {